# Improve font performance in FamilyCollection.GetFontFromFamily